+++ packages/ai/core/generate-text/generate-text.ts
@@ -88,9 +88,7 @@ export async function generateText<TOOLS extends Record<string, CoreTool>>({
   abortSignal,
   headers,
   maxSteps = 1,
-  experimental_continuationSteps,
-  experimental_continueSteps: continueSteps = experimental_continuationSteps ??
-    false,
+  experimental_continueSteps: continueSteps = false,
   experimental_telemetry: telemetry,
   experimental_providerMetadata: providerMetadata,
   experimental_activeTools: activeTools,
@@ -125,11 +123,6 @@ A maximum number is required to prevent infinite loops in the case of misconfigu
 By default, it's set to 1, which means that only a single LLM call is made.
      */
     maxSteps?: number;
-
-    /**
-@deprecated Use `experimental_continueSteps` instead.
-     */
-    experimental_continuationSteps?: boolean;
 
     /**
 When enabled, the model will perform additional steps if the finish reason is "length" (experimental).
@@ -498,7 +491,6 @@ changing the tool call and result types in the result.
           messages: responseMessages,
         },
         logprobs: currentModelResponse.logprobs,
-        responseMessages,
         steps,
         providerMetadata: currentModelResponse.providerMetadata,
       });
@@ -592,7 +584,6 @@ class DefaultGenerateTextResult<TOOLS extends Record<string, CoreTool>>
   readonly finishReason: GenerateTextResult<TOOLS>['finishReason'];
   readonly usage: GenerateTextResult<TOOLS>['usage'];
   readonly warnings: GenerateTextResult<TOOLS>['warnings'];
-  readonly responseMessages: GenerateTextResult<TOOLS>['responseMessages'];
   readonly steps: GenerateTextResult<TOOLS>['steps'];
   readonly logprobs: GenerateTextResult<TOOLS>['logprobs'];
   readonly experimental_providerMetadata: GenerateTextResult<TOOLS>['experimental_providerMetadata'];
@@ -607,7 +598,6 @@ class DefaultGenerateTextResult<TOOLS extends Record<string, CoreTool>>
     usage: GenerateTextResult<TOOLS>['usage'];
     warnings: GenerateTextResult<TOOLS>['warnings'];
     logprobs: GenerateTextResult<TOOLS>['logprobs'];
-    responseMessages: GenerateTextResult<TOOLS>['responseMessages'];
     steps: GenerateTextResult<TOOLS>['steps'];
     providerMetadata: GenerateTextResult<TOOLS>['experimental_providerMetadata'];
     response: GenerateTextResult<TOOLS>['response'];
@@ -621,7 +611,6 @@ class DefaultGenerateTextResult<TOOLS extends Record<string, CoreTool>>
     this.warnings = options.warnings;
     this.request = options.request;
     this.response = options.response;
-    this.responseMessages = options.responseMessages;
     this.steps = options.steps;
     this.experimental_providerMetadata = options.providerMetadata;
     this.logprobs = options.logprobs;