+++ packages/ai/core/types/index.ts
@@ -1,32 +1,17 @@
-import type {
-  LanguageModelUsage as LanguageModelUsageOriginal,
-  EmbeddingModelUsage as EmbeddingModelUsageOriginal,
-} from './usage';
-
-export * from './embedding-model';
-export * from './language-model';
-
-export type { LanguageModelRequestMetadata } from './language-model-request-metadata';
+export type { Embedding, EmbeddingModel } from './embedding-model';
 export type {
-  LanguageModelResponseMetadata,
-  LanguageModelResponseMetadataWithHeaders,
-} from './language-model-response-metadata';
-
+  CallWarning,
+  CoreToolChoice,
+  FinishReason,
+  LanguageModel,
+  LanguageModelV1,
+  LanguageModelV1CallOptions,
+  LanguageModelV1Prompt,
+  LanguageModelV1StreamPart,
+  LogProbs,
+} from './language-model';
+export type { LanguageModelRequestMetadata } from './language-model-request-metadata';
+export type { LanguageModelResponseMetadata } from './language-model-response-metadata';
 export type { Provider } from './provider';
 export type { ProviderMetadata } from './provider-metadata';
-
-/**
- * @deprecated Use LanguageModelUsage instead.
- */
-export type TokenUsage = LanguageModelUsageOriginal;
-/**
- * @deprecated Use LanguageModelUsage instead.
- */
-export type CompletionTokenUsage = LanguageModelUsageOriginal;
-export type LanguageModelUsage = LanguageModelUsageOriginal;
-
-/**
- * @deprecated Use EmbeddingModelUsage instead.
- */
-export type EmbeddingTokenUsage = EmbeddingModelUsageOriginal;
-export type EmbeddingModelUsage = EmbeddingModelUsageOriginal;
+export type { EmbeddingModelUsage, LanguageModelUsage } from './usage';